add checkpoints option to the param json
