+++ tierpsy/extras/param_files/_TEST_WT2.json
@@ -1,0 +1,18 @@
+{
+	"thresh_block_size": 61, 
+	"thresh_C": 15, 
+	"dilation_size": 10, 
+	"min_area": 100, 
+	"resampling_N": 49, 
+	"expected_fps": 30.0, 
+	"max_area": 100000, 
+	"is_extract_metadata": true,
+	"keep_border_data": true, 
+	"compression_buff": 1, 
+	"worm_bw_thresh_factor":1.0, 
+	"strel_size":10, 
+	"int_save_maps": true, 
+	"filt_min_displacement":0,
+	"analysis_type" : "WT2",
+	"analysis_checkpoints" : ["FEAT_INIT", "FEAT_TIERPSY"]
+}
